renamed .secure to .security as per Max suggestion - JBIDE-15496
