+++ plugins/org.jboss.tools.openshift.express.core/src/org/jboss/tools/openshift/express/internal/core/behaviour/OpenshiftDelegatingStartLaunchConfiguration.java
@@ -16,6 +16,7 @@ import org.jboss.ide.eclipse.as.core.server.internal.launch.DelegatingStartLaunc
 /**
  * @author Rob Stryker
  */
-public class OpenshiftDelegatingStartLaunchConfiguration extends DelegatingStartLaunchConfiguration {
+public class OpenshiftDelegatingStartLaunchConfiguration 
+	extends DelegatingStartLaunchConfiguration implements ILaunchConfigurationDelegate {
 
 }
