+++ plugins/org.jboss.tools.openshift.express.core/src/org/jboss/tools/openshift/express/internal/core/security/SecurePasswordStore.java
@@ -8,7 +8,7 @@
  * Contributors:
  *     Red Hat, Inc. - initial API and implementation
  ******************************************************************************/
-package org.jboss.tools.openshift.express.internal.core.secure;
+package org.jboss.tools.openshift.express.internal.core.security;
 
 import java.io.UnsupportedEncodingException;
